+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraRequestExceptionHandler.java
@@ -145,7 +145,9 @@ class CassandraRequestExceptionHandler {
 
     @VisibleForTesting
     boolean shouldBlacklist(Exception ex, int numberOfAttempts) {
-        return isConnectionException(ex) && numberOfAttempts >= maxTriesSameHost.get();
+        return isConnectionException(ex)
+                && numberOfAttempts >= maxTriesSameHost.get()
+                && !isExceptionNotImplicatingThisParticularNode(ex);
     }
 
     private <K extends Exception> void handleBackoff(RetryableCassandraRequest<?, K> req,
@@ -234,6 +236,15 @@ class CassandraRequestExceptionHandler {
                 // underlying cassandra table does not exist. The table might exist on other cassandra nodes.
                 && (ex instanceof InvalidRequestException
                 || isFastFailoverException(ex.getCause()));
+    }
+
+    static boolean isExceptionNotImplicatingThisParticularNode(Throwable ex) {
+        // client pool has no more elements to give, so this service node has too many requests in flight.
+        return ex instanceof NoSuchElementException
+                // Other nodes are down. Technically this node might be isolated from a functional quorum via a network
+                // partition but we consider that to be an edge case.
+                || ex instanceof InsufficientConsistencyException
+                || isIndicativeOfCassandraLoad(ex.getCause());
     }
 
     @VisibleForTesting